this

+++ controllers/Category.js
@@ -1,6 +1,4 @@
 const Category = require("../models/productCategory");
-const Product = require("../models/productModel");
-const { ObjectId } = require('mongodb');
 
 const { uploadToCloudinary } = require("../utils/imageUploader");
 
@@ -54,13 +52,20 @@ exports.createCategory = async (req, res) => {
 exports.showAllCategory = async (req, res) => {
   try {
 
-    const allCategory = await Category.find({}, { title: true , images:true });
-
+    const allCategory = await Category.find({}, { title: true, images: true, subCategory: true })
+    .populate({
+        path: "subCategory",
+        populate: {
+            path: "products", 
+            model: "Product", 
+        },
+    });
     res.status(200).json({
       success: true,
       message: "all category return succesfully",
      data:allCategory,
     });
+
   } catch (error) {
     return res.status(500).json({
       success: false,
@@ -69,67 +74,15 @@ exports.showAllCategory = async (req, res) => {
   }
 };
 
-// ! category page details -> see ui different different types ke course show krna 
-exports.categoryPageDetails = async(req,res)=>{
-  try{
-
-    // get category id
-    const {categoryId} = req.params;
-
-    // get courses for specified category id
-    const selectedCategory = await Category.findById(categoryId).populate({
-      path:"products",
-    }).exec(); 
-
-     console.log("seletedCategory" , selectedCategory);
-     
-
-    // validation 
-    if(!selectedCategory){
-      return res.status(404).json({
-        success:false,
-        message:"categoy not found "
-      })
-    }
-
-    if(selectedCategory.products.length === 0){
-      return res.json({
-        success:false,
-        message:"no product found for the selected category",
-      })
-    }
-   
-
-    // ṛeturn 
-    return res.status(200).json({
-      success:true,
-      message:"successfuly fetch all category details ",
-    data:{
-      selectedCategory,
-    
-    }
-    })
-         
-  } catch(error){
- console.log(error);
- return res.status(500).json({
-  success:false,
-  message:error.message
- })
-  }
-}
-
 // ! delete category 
 exports.deleteCategory = async(req , res)=>{
   try{
 
     const {categoryId} = req.params;
 
-    console.log("categoryiD" ,categoryId);
 
-     const categoryDetail = await Category.findOne({_id:categoryId});
+     const categoryDetail = await Category.findOne({_id:categoryId}).populate("subCategory");
 
-     console.log("categoryDe" , categoryDetail);
 
      if(!categoryDetail){
       return res.status(404).json({
@@ -138,15 +91,13 @@ exports.deleteCategory = async(req , res)=>{
       })
      }
 
-   for (let index = 0; index < categoryDetail.products.length; index++) {
-    const productId = categoryDetail.products[index];
-
-    const productIdObjectId = new ObjectId(productId);
-
-
-     await Product.findByIdAndDelete({_id:productIdObjectId} , {new:true});
-    
-   }
+   
+    if(categoryDetail.subCategory.length > 0 ){
+      return res.status(403).json({
+        success:false ,
+        message:"Please Delete All the Sub Category"
+      })
+    }
 
      await Category.findByIdAndDelete({_id:categoryId} , {new:true});
